Move instrument method to correct location.
