+++ tests/mobly/controllers/android_device_lib/adb_test.py
@@ -14,9 +14,31 @@
 
 import mock
 
+from collections import OrderedDict
 from future.tests.base import unittest
 
 from mobly.controllers.android_device_lib import adb
+
+# Mock parameters for instrumentation.
+MOCK_INSTRUMENTATION_PACKAGE = 'com.my.instrumentation.tests'
+MOCK_INSTRUMENTATION_RUNNER = 'com.my.instrumentation.runner'
+MOCK_INSTRUMENTATION_OPTIONS = OrderedDict([
+    ('option1', 'value1'),
+    ('option2', 'value2'),
+])
+# Mock android instrumentation commands.
+MOCK_BASIC_INSTRUMENTATION_COMMAND = ('am instrument -r -w  com.my'
+                                      '.instrumentation.tests/com.android'
+                                      '.common.support.test.runner'
+                                      '.AndroidJUnitRunner')
+MOCK_RUNNER_INSTRUMENTATION_COMMAND = ('am instrument -r -w  com.my'
+                                       '.instrumentation.tests/com.my'
+                                       '.instrumentation.runner')
+MOCK_OPTIONS_INSTRUMENTATION_COMMAND = ('am instrument -r -w -e option1 value1'
+                                        ' -e option2 value2 com.my'
+                                        '.instrumentation.tests/com.android'
+                                        '.common.support.test.runner'
+                                        '.AndroidJUnitRunner')
 
 
 class AdbTest(unittest.TestCase):
@@ -31,8 +53,8 @@ class AdbTest(unittest.TestCase):
         # the created process object in adb._exec_cmd()
         mock_psutil_process.return_value = mock.Mock()
 
-        mock_proc.communicate = mock.Mock(return_value=("out".encode('utf-8'),
-                                                        "err".encode('utf-8')))
+        mock_proc.communicate = mock.Mock(
+            return_value=('out'.encode('utf-8'), 'err'.encode('utf-8')))
         mock_proc.returncode = 0
         return (mock_psutil_process, mock_popen)
 
@@ -43,8 +65,8 @@ class AdbTest(unittest.TestCase):
         self._mock_process(mock_psutil_process, mock_Popen)
 
         reply = adb.AdbProxy()._exec_cmd(
-            ["fake_cmd"], shell=False, timeout=None)
-        self.assertEqual("out", reply.decode('utf-8'))
+            ['fake_cmd'], shell=False, timeout=None)
+        self.assertEqual('out', reply.decode('utf-8'))
 
     @mock.patch('mobly.controllers.android_device_lib.adb.subprocess.Popen')
     @mock.patch('mobly.controllers.android_device_lib.adb.psutil.Process')
@@ -54,8 +76,8 @@ class AdbTest(unittest.TestCase):
         mock_popen.return_value.returncode = 1
 
         with self.assertRaisesRegex(adb.AdbError,
-                                    "Error executing adb cmd .*"):
-            adb.AdbProxy()._exec_cmd(["fake_cmd"], shell=False, timeout=None)
+                                    'Error executing adb cmd .*'):
+            adb.AdbProxy()._exec_cmd(['fake_cmd'], shell=False, timeout=None)
 
     @mock.patch('mobly.controllers.android_device_lib.adb.subprocess.Popen')
     @mock.patch('mobly.controllers.android_device_lib.adb.psutil.Process')
@@ -63,8 +85,8 @@ class AdbTest(unittest.TestCase):
                                            mock_popen):
         self._mock_process(mock_psutil_process, mock_popen)
 
-        reply = adb.AdbProxy()._exec_cmd(["fake_cmd"], shell=False, timeout=1)
-        self.assertEqual("out", reply.decode('utf-8'))
+        reply = adb.AdbProxy()._exec_cmd(['fake_cmd'], shell=False, timeout=1)
+        self.assertEqual('out', reply.decode('utf-8'))
 
     @mock.patch('mobly.controllers.android_device_lib.adb.subprocess.Popen')
     @mock.patch('mobly.controllers.android_device_lib.adb.psutil.Process')
@@ -76,8 +98,8 @@ class AdbTest(unittest.TestCase):
             adb.psutil.TimeoutExpired('Timed out'))
 
         with self.assertRaisesRegex(adb.AdbTimeoutError,
-                                    "Timed out Adb cmd .*"):
-            adb.AdbProxy()._exec_cmd(["fake_cmd"], shell=False, timeout=0.1)
+                                    'Timed out Adb cmd .*'):
+            adb.AdbProxy()._exec_cmd(['fake_cmd'], shell=False, timeout=0.1)
 
     @mock.patch('mobly.controllers.android_device_lib.adb.subprocess.Popen')
     @mock.patch('mobly.controllers.android_device_lib.adb.psutil.Process')
@@ -85,8 +107,8 @@ class AdbTest(unittest.TestCase):
                                                   mock_popen):
         self._mock_process(mock_psutil_process, mock_popen)
         with self.assertRaisesRegex(adb.AdbError,
-                                    "Timeout is a negative value: .*"):
-            adb.AdbProxy()._exec_cmd(["fake_cmd"], shell=False, timeout=-1)
+                                    'Timeout is a negative value: .*'):
+            adb.AdbProxy()._exec_cmd(['fake_cmd'], shell=False, timeout=-1)
 
     def test_exec_adb_cmd(self):
         with mock.patch.object(adb.AdbProxy, '_exec_cmd') as mock_exec_cmd:
@@ -110,6 +132,43 @@ class AdbTest(unittest.TestCase):
             mock_exec_cmd.assert_called_once_with(
                 '"adb" -s "12345" shell arg1 arg2', shell=True, timeout=None)
 
+    def test_AndroidDevice_instrument_without_parameters(self):
+        """Verifies the AndroidDevice object's instrument command is correct in
+        the basic case.
+        """
+        with mock.patch.object(adb.AdbProxy, '_exec_cmd') as mock_exec_cmd:
+            adb.AdbProxy().instrument(MOCK_INSTRUMENTATION_PACKAGE)
+            mock_exec_cmd.assert_called_once_with(
+                ['adb', 'shell', MOCK_BASIC_INSTRUMENTATION_COMMAND],
+                shell=False,
+                timeout=None)
 
-if __name__ == "__main__":
+    def test_AndroidDevice_instrument_with_runner(self):
+        """Verifies the AndroidDevice object's instrument command is correct
+        with a runner specified.
+        """
+        with mock.patch.object(adb.AdbProxy, '_exec_cmd') as mock_exec_cmd:
+            adb.AdbProxy().instrument(
+                MOCK_INSTRUMENTATION_PACKAGE,
+                runner=MOCK_INSTRUMENTATION_RUNNER)
+            mock_exec_cmd.assert_called_once_with(
+                ['adb', 'shell', MOCK_RUNNER_INSTRUMENTATION_COMMAND],
+                shell=False,
+                timeout=None)
+
+    def test_AndroidDevice_instrument_with_options(self):
+        """Verifies the AndroidDevice object's instrument command is correct
+        with options.
+        """
+        with mock.patch.object(adb.AdbProxy, '_exec_cmd') as mock_exec_cmd:
+            adb.AdbProxy().instrument(
+                MOCK_INSTRUMENTATION_PACKAGE,
+                options=MOCK_INSTRUMENTATION_OPTIONS)
+            mock_exec_cmd.assert_called_once_with(
+                ['adb', 'shell', MOCK_OPTIONS_INSTRUMENTATION_COMMAND],
+                shell=False,
+                timeout=None)
+
+
+if __name__ == '__main__':
     unittest.main()
